Add tokenizer types
